Extract signatures classes into separate module
