+++ tests/fixtures/models.py
@@ -4,7 +4,7 @@ from typing import (Any,
 
 import pytest
 
-from paradigm import signatures
+from paradigm import models
 from tests import strategies
 from tests.utils import (find,
                          is_signature_empty,
@@ -12,66 +12,66 @@ from tests.utils import (find,
 
 
 @pytest.fixture(scope='function')
-def signature() -> signatures.Base:
+def signature() -> models.Base:
     return find(strategies.signatures)
 
 
 @pytest.fixture(scope='function')
-def other_signature() -> signatures.Base:
+def other_signature() -> models.Base:
     return find(strategies.signatures)
 
 
 @pytest.fixture(scope='function')
-def another_signature() -> signatures.Base:
+def another_signature() -> models.Base:
     return find(strategies.signatures)
 
 
 @pytest.fixture(scope='function')
-def overloaded_signature() -> signatures.Overloaded:
+def overloaded_signature() -> models.Overloaded:
     return find(strategies.overloaded_signatures)
 
 
 @pytest.fixture(scope='function')
-def non_empty_signature() -> signatures.Base:
+def non_empty_signature() -> models.Base:
     return find(strategies.signatures.filter(negate(is_signature_empty)))
 
 
 @pytest.fixture(scope='function')
-def non_variadic_signature() -> signatures.Base:
+def non_variadic_signature() -> models.Base:
     return find(strategies.non_variadic_signatures)
 
 
 @pytest.fixture(scope='function')
-def non_empty_signature_expected_kwargs(non_empty_signature: signatures.Base
+def non_empty_signature_expected_kwargs(non_empty_signature: models.Base
                                         ) -> Dict[str, Any]:
     return find(strategies.to_expected_kwargs(non_empty_signature))
 
 
 @pytest.fixture(scope='function')
-def non_empty_signature_expected_args(non_empty_signature: signatures.Base
+def non_empty_signature_expected_args(non_empty_signature: models.Base
                                       ) -> Tuple[Any, ...]:
     return find(strategies.to_expected_args(non_empty_signature))
 
 
 @pytest.fixture(scope='function')
 def non_variadic_signature_expected_args(
-        non_variadic_signature: signatures.Base) -> Tuple[Any, ...]:
+        non_variadic_signature: models.Base) -> Tuple[Any, ...]:
     return find(strategies.to_expected_args(non_variadic_signature))
 
 
 @pytest.fixture(scope='function')
 def non_variadic_signature_expected_kwargs(
-        non_variadic_signature: signatures.Base) -> Dict[str, Any]:
+        non_variadic_signature: models.Base) -> Dict[str, Any]:
     return find(strategies.to_expected_kwargs(non_variadic_signature))
 
 
 @pytest.fixture(scope='function')
 def non_variadic_signature_unexpected_args(
-        non_variadic_signature: signatures.Base) -> Tuple[Any, ...]:
+        non_variadic_signature: models.Base) -> Tuple[Any, ...]:
     return find(strategies.to_unexpected_args(non_variadic_signature))
 
 
 @pytest.fixture(scope='function')
 def non_variadic_signature_unexpected_kwargs(
-        non_variadic_signature: signatures.Base) -> Dict[str, Any]:
+        non_variadic_signature: models.Base) -> Dict[str, Any]:
     return find(strategies.to_unexpected_kwargs(non_variadic_signature))
