+++ tests/models_tests/test_bind.py
@@ -7,17 +7,17 @@ from typing import (Any,
 
 import pytest
 
-from paradigm import signatures
+from paradigm import models
 from tests.utils import implication
 
 
-def test_basic(non_empty_signature: signatures.Base) -> None:
+def test_basic(non_empty_signature: models.Base) -> None:
     result = non_empty_signature.bind()
 
     assert result == non_empty_signature
 
 
-def test_expected_args(non_empty_signature: signatures.Base,
+def test_expected_args(non_empty_signature: models.Base,
                        non_empty_signature_expected_args: Tuple[Any, ...]
                        ) -> None:
     result = non_empty_signature.bind(*non_empty_signature_expected_args)
@@ -26,7 +26,7 @@ def test_expected_args(non_empty_signature: signatures.Base,
                        result != non_empty_signature)
 
 
-def test_expected_kwargs(non_empty_signature: signatures.Base,
+def test_expected_kwargs(non_empty_signature: models.Base,
                          non_empty_signature_expected_kwargs: Dict[str, Any]
                          ) -> None:
     result = non_empty_signature.bind(**non_empty_signature_expected_kwargs)
@@ -38,24 +38,24 @@ def test_expected_kwargs(non_empty_signature: signatures.Base,
 
 
 @singledispatch
-def signature_parameters_has_defaults(signature: signatures.Base,
+def signature_parameters_has_defaults(signature: models.Base,
                                       *,
                                       names: Iterable[str]) -> bool:
     raise TypeError('Unsupported signature type: {type}.'
                     .format(type=type(signature)))
 
 
-@signature_parameters_has_defaults.register(signatures.Plain)
-def plain_signature_parameters_has_defaults(signature: signatures.Plain,
+@signature_parameters_has_defaults.register(models.Plain)
+def plain_signature_parameters_has_defaults(signature: models.Plain,
                                             *,
                                             names: Iterable[str]) -> bool:
-    parameters = signatures.to_parameters_by_name(signature.parameters)
+    parameters = models.to_parameters_by_name(signature.parameters)
     return all(parameters[name].has_default for name in names)
 
 
-@signature_parameters_has_defaults.register(signatures.Overloaded)
+@signature_parameters_has_defaults.register(models.Overloaded)
 def overloaded_signature_parameters_has_defaults(
-        signature: signatures.Overloaded,
+        signature: models.Overloaded,
         *,
         names: Iterable[str]) -> bool:
     return all(map(partial(signature_parameters_has_defaults,
@@ -64,14 +64,14 @@ def overloaded_signature_parameters_has_defaults(
 
 
 def test_unexpected_args(
-        non_variadic_signature: signatures.Base,
+        non_variadic_signature: models.Base,
         non_variadic_signature_unexpected_args: Tuple[Any, ...]) -> None:
     with pytest.raises(TypeError):
         non_variadic_signature.bind(*non_variadic_signature_unexpected_args)
 
 
 def test_unexpected_kwargs(
-        non_variadic_signature: signatures.Base,
+        non_variadic_signature: models.Base,
         non_variadic_signature_unexpected_kwargs: Dict[str, Any]) -> None:
     with pytest.raises(TypeError):
         non_variadic_signature.bind(**non_variadic_signature_unexpected_kwargs)
