+++ tests/models_tests/test_str.py
@@ -1,13 +1,13 @@
-from paradigm import signatures
+from paradigm import models
 
 
-def test_type(signature: signatures.Base) -> None:
+def test_type(signature: models.Base) -> None:
     result = str(signature)
 
     assert isinstance(result, str)
 
 
-def test_nesting(overloaded_signature: signatures.Overloaded) -> None:
+def test_nesting(overloaded_signature: models.Overloaded) -> None:
     result = str(overloaded_signature)
 
     assert all(str(signature) in result
