+++ tests/strategies/__init__.py
@@ -9,13 +9,13 @@ from .definitions import (built_in_functions,
                           partial_callables,
                           unsupported_callables,
                           wrappers_descriptors)
-from .signatures.base import (non_variadic_signatures,
-                              overloaded_signatures,
-                              signatures)
-from .signatures.factories import (to_expected_args,
-                                   to_expected_kwargs,
-                                   to_unexpected_args,
-                                   to_unexpected_kwargs)
+from .models.base import (non_variadic_signatures,
+                          overloaded_signatures,
+                          signatures)
+from .models.factories import (to_expected_args,
+                               to_expected_kwargs,
+                               to_unexpected_args,
+                               to_unexpected_kwargs)
 from .utils import (identifiers,
                     invalid_sources,
                     paths)
